src/requirements.txt
fastapi
uvicorn
requests
pydantic
pyyaml
autoagents-core
